# src/main/resources/v1/eric-oss-auto-config-consistency-openapi.yaml
...
openapi: "3.0.2"
servers:
  - url: http://eacc/v1
info:
  x-api-id: 8608e72f-121d-4049-be33-7eb2d3cf5594
  x-audience: external-public
  description: |
    Ericsson Automated Configuration Consistency (EACC) provides self-service analytics 
    to enable users to uncover baseline consistency errors. EACC supports the scheduling 
    of CM consistency audits which detect inconsistencies based on user-supplied rules. 
    These inconsistencies can be corrected manually by the user. 
    Any changes can also be manually rolled back by the user.
  version: 0.1.5
  title: Ericsson Automated Configuration Consistency
  license:
    name: COPYRIGHT Ericsson 2024
    url: https://www.ericsson.com/en/legal
  termsOfService: https://www.ericsson.com/en/legal
  contact:
    email: pdlonteamq@pdl.internal.ericsson.com
...
paths:
...
  /v1/jobs:
    get:
      tags:
        - jobs
      summary: Return all the Jobs
      description: Get all the Jobs defined in EACC
      operationId: getJobs
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Job'
              example:
                [
                  {
                    "jobName": "job_market1",
                    "scopeName": "market1",
                    "schedule": "0 38 14 19 2 *",
                    "rulesetName": "ruleset1"
                  },
                  {
                    "jobName": "job_market2",
                    "scopeName": "market2",
                    "schedule": "0 50 6 20 2 *",
                    "rulesetName": "ruleset2"
                  }
                ]
        default:
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
              example: {
                "title": "Failed to get job(s).",
                "status": 500,
                "detail": "Database operation failed."
              }
    post:
      tags:
        - jobs
      summary: Create a Job
      description: Create a Job in EACC defining the schedule, rules and Scope of the configuration consistency check.
      operationId: postJobs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Job'
            example:
              {
                "jobName": "job_market1",
                "scopeName": "market1",
                "schedule": "0 38 14 19 2 *",
                "rulesetName": "ruleset1"
              }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
              example:
                {
                  "jobName": "job_market1",
                  "scopeName": "market1",
                  "schedule": "0 38 14 19 2 *",
                  "rulesetName": "ruleset1"
                }
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
              example: {
                "title": "Validation failed.",
                "status": 400,
                "detail": "Invalid jobName: job already exists"
              }
        default:
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
              example: {
                "title": "Failed to create job.",
                "status": 500,
                "detail": "Database operation failed."
              }
  /v1/jobs/{jobName}:
    put:
      tags:
        - jobs
      summary: Update a Job
      description: |
        Update a Job in EACC to change the schedule, rules or Scope of the configuration consistency check.
        The Job name cannot be modified and if it is included in the payload it will be ignored.
        A Job cannot be updated if there is an ongoing execution involving the Job.
      operationId: putJob
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Job'
            example:
              {
                "jobName": "job_market1",
                "scopeName": "market1",
                "schedule": "0 38 15 19 2 *",
                "rulesetName": "ruleset1"
              }
      parameters:
        - $ref: '#/components/parameters/acceptInHeader'
        - $ref: '#/components/parameters/contentTypeInHeader'
        - $ref: '#/components/parameters/jobName'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
              example:
                {
                  "jobName": "job_market1",
                  "scopeName": "market1",
                  "schedule": "0 38 15 19 2 *",
                  "rulesetName": "ruleset1"
                }
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
              example: {
                "title": "Validation failed",
                "status": 400,
                "detail": "Invalid resource identifier in url: Only lower case alphanumeric, underscores and dashes allowed, with between 4 - 100 characters."
              }
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
              example: {
                "title": "Validation failed",
                "status": 404,
                "detail": "Invalid jobName: jobName does not exist"
              }
        "409":
          description: Conflict
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
              example: {
                "title": "Cannot update job as it is used by an ongoing execution.",
                "status": 409,
                "detail": "Job in use."
              }
        default:
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
              example: {
                "title": "Failed to update job.",
                "status": 500,
                "detail": "Database operation failed."
              }
    delete:
      tags:
        - jobs
      summary: Delete a Job.
      description: Delete a Job in EACC.
      operationId: deleteJob
      parameters:
        - $ref: '#/components/parameters/acceptInHeader'
        - $ref: '#/components/parameters/jobName'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
              example: {
                "title": "Validation failed",
                "status": 400,
                "detail": "Invalid resource identifier in url: Only lower case alphanumeric, underscores and dashes allowed, with between 4 -100 characters."
              }
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
              example: {
                "title": "Validation failed",
                "status": 404,
                "detail": "Invalid jobName: jobName does not exist"
              }
        default:
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
              example: {
                "title": "Failed to delete job.",
                "status": 500,
                "detail": "Database operation failed."
              }
...
components:
...
  parameters:
    jobName:
      name: jobName
      in: path
      description: Name of the job
      required: true
      schema:
        type: string
...
    acceptInHeader:
      name: Accept
      in: header
      required: false
      schema:
        type: string
        example: application/json
        default: application/json
    contentTypeInHeader:
      name: Content-Type
      in: header
      required: false
      schema:
        type: string
        example: application/json
        default: application/json

  # Common error responses
...
  schemas:
...
    Job:
      type: object
      required: [ jobName, scopeName, schedule, rulesetName ]
      properties:
        jobName:
          type: string
          description: The name of the Job.
        scopeName:
          type: string
          description: The Scope to be used in the Job.
        schedule:
          type: string
          description: The schedule defined as a cron expression
        rulesetName:
          type: string
          description: The Ruleset to be used in the Job.
...
    ErrorDetails:
      description: Provides additional information in an error response.
      type: object
      required: [ title, status, detail ]
      properties:
        title:
          $ref: '#/components/schemas/ErrorTitle'
        status:
          $ref: '#/components/schemas/ErrorStatus'
        detail:
          $ref: '#/components/schemas/ErrorDetail'
...
    ErrorTitle:
      type: string
      description: Short, human-readable summary of the problem.
    ErrorStatus:
      type: integer
      description: HTTP status code.
    ErrorDetail:
      description: |
        Optional human-readable explanation of the problem. if present, ought to focus on helping the client 
        correct the problem, rather than giving debugging information.